+++ nodeconductor/cost_tracking/views.py
@@ -34,6 +34,11 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
     filter_class = filters.PriceEstimateFilter
     permission_classes = (permissions.IsAuthenticated,)
 
+    def get_serializer_class(self):
+        if self.action == 'threshold':
+            return serializers.PriceEstimateThresholdSerializer
+        return self.serializer_class
+
     def get_queryset(self):
         return models.PriceEstimate.objects.filtered_for_user(self.request.user).filter(is_visible=True).order_by(
             '-year', '-month')
@@ -67,7 +72,7 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
 
         Request example:
 
-        .. code-block:: javascript
+        .. code-block:: http
 
             POST /api/price-estimates/
             Accept: application/json
@@ -102,13 +107,27 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
         """
         Run **POST** request against */api/price-estimates/threshold/*
         to set alert threshold for price estimate.
-        Request body should be JSON dictionary with scope and threshold fields.
+        Example request:
+
+        .. code-block:: http
+
+            POST /api/price-estimates/threshold/
+            Accept: application/json
+            Content-Type: application/json
+            Authorization: Token c84d653b9ec92c6cbac41c706593e66f567a7fa4
+            Host: example.com
+
+            {
+                "scope": "http://example.com/api/projects/ab2e3d458e8a4ecb9dded36f3e46878d/",
+                "threshold": 100.0,
+                "limit": 100.0
+            }
         """
-        serializer = serializers.PriceEstimateThresholdSerializer(
-            data=request.data, context={'request': request})
+        serializer = self.get_serializer(data=request.data)
         serializer.is_valid(raise_exception=True)
 
         threshold = serializer.validated_data['threshold']
+        limit = serializer.validated_data['limit']
         scope = serializer.validated_data['scope']
 
         if not self.can_user_modify_price_object(scope):
@@ -117,9 +136,9 @@ class PriceEstimateViewSet(PriceEditPermissionMixin, viewsets.ModelViewSet):
         today = timezone.now()
         params = dict(scope=scope, year=today.year, month=today.month)
         try:
-            models.PriceEstimate.objects.create(threshold=threshold, **params)
+            models.PriceEstimate.objects.create(threshold=threshold, limit=limit, **params)
         except IntegrityError:
-            models.PriceEstimate.objects.filter(**params).update(threshold=threshold)
+            models.PriceEstimate.objects.filter(**params).update(threshold=threshold, limit=limit)
 
         return response.Response({'detail': 'Threshold for price estimate is updated'},
                                  status=status.HTTP_200_OK)